+++ src/layouts/center_main.rs
@@ -69,7 +69,7 @@ pub fn update(workspace: &Workspace, windows: &mut Vec<&mut &mut Window>) {
     };
 
     let primary_x = match window_count {
-        1 => 0 as i32,
+        1 => 0_i32,
         2 => (workspace.width() as f32 / 2.0).floor() as i32,
         _ => (workspace.width() as f32 / 4.0).floor() as i32,
     };
